pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "breakagent"
version = "0.1.0"
description = "Adaptive API stress-testing agent for security, robustness, and quality"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1.7",
  "rich>=13.7.1",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "Jinja2>=3.1.4",
  "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
breakagent = "breakagent.cli:cli"

[tool.setuptools]
packages = ["breakagent", "breakagent.modules", "breakagent.planners"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
